+++ src/Components/Alert/PopUp/ExitApp.tsx
@@ -1,15 +1,13 @@
 import React, { useEffect, useMemo } from 'react';
 import { View } from 'react-native';
-import * as Vibration from 'expo-haptics';
 
 import ConfigService from '@Services/ConfigService';
 import AlertService from '@Services/AlertService';
 import ThemeService from '@Services/ThemeService';
+import ApticsService from '@Services/ApticsService';
 
 import { Text } from '@Text/index';
 import { Button } from '@Button/index';
-
-type Vibration = 'warning' | 'success'
 
 export default function ExitApp(props: {
   question: string
@@ -21,18 +19,18 @@ export default function ExitApp(props: {
   const theme  = useMemo(() => ThemeService.appThemes[config.appTheme].layout.modalPopUp, []);
 
   useEffect(() => {
-    vibrate('warning');
+    ApticsService.vibrate('warning');
   }, []);
 
   async function onAccept() {
     props.onAccept();
     await AlertService.runAcceptCallback();
-    await vibrate('warning');
+    ApticsService.vibrate('warning');
   }
 
   async function onRefuse() {
     props.onRefuse();
-    await vibrate('success');
+    ApticsService.vibrate('success');
   }
 
   return (
@@ -104,11 +102,4 @@ export default function ExitApp(props: {
       </View>
     </View>
   );
-}
-
-async function vibrate(type: Vibration) {
-  switch (type) {
-    case 'warning': await Vibration.notificationAsync(Vibration.NotificationFeedbackType.Warning); break;
-    case 'success': await Vibration.notificationAsync(Vibration.NotificationFeedbackType.Success); break;
-  }
 }