+++ src/Screens/SettingsScreen/index.tsx
@@ -1,7 +1,6 @@
 
 import React, { useMemo } from 'react';
-import { Alert, View } from 'react-native';
-import * as Vibration from 'expo-haptics';
+import { View } from 'react-native';
 
 import { navigate } from '@Globals/NavigationControler';
 import { translations } from '@Translations/index';
@@ -13,6 +12,8 @@ import ThemeService from '@Services/ThemeService';
 import { Button } from '@Button/index';
 import { Layout } from '@Layout/index';
 import { TC } from './__TC__';
+import ApticsService from '@Services/ApticsService';
+import AlertService from '@Services/AlertService';
 
 export default function SettingsScreen(): JSX.Element {
 
@@ -25,9 +26,7 @@ export default function SettingsScreen(): JSX.Element {
       screenButtons={<TC.ScreenButtons />}
     >
       <View
-        style={{
-          gap: 1,
-        }}
+        style={{ gap: 1 }}
       >
         <Button.TextWithIcon
           title={R['Language']}
@@ -61,43 +60,31 @@ export default function SettingsScreen(): JSX.Element {
 async function whipeDataBase() {
 
   const R = translations.Screens.SettingsScreen[ConfigService.config.language];
-  await Vibration.notificationAsync(Vibration.NotificationFeedbackType.Warning);
 
-  Alert.alert(
-    R['Hold on!'],
-    R['Want to whipe database?'],
-    [
-      {
-        text: R['NO'],
-        onPress: async () => {
-          await Vibration.notificationAsync(Vibration.NotificationFeedbackType.Success);
-          return null;
+  AlertService.handleAlert(true,
+    {
+      question: R['Want to whipe database?'],
+      type: 'warning',
+    },
+    async () => {
+      ApticsService.vibrate('success');
+      await DatabaseService.deleteDatabase();
+      await CacheService.deleteLastOpenProject();
+      CacheService.lastOpenProject = {
+        id_project: '',
+        name: '',
+        rules: {},
+        sampleAlias: {
+          plural: '',
+          singular: '',
         },
-        style: 'cancel',
-      },
-      {
-        text: R['YES'],
-        onPress: async () => {
-          await Vibration.notificationAsync(Vibration.NotificationFeedbackType.Success);
-          await DatabaseService.deleteDatabase();
-          await CacheService.deleteLastOpenProject();
-          CacheService.lastOpenProject = {
-            id_project: '',
-            name: '',
-            rules: {},
-            sampleAlias: {
-              plural: '',
-              singular: '',
-            },
-          };
-          CacheService.allProjects = [];
-          CacheService.allWidgets_Project = [];
-          CacheService.allWidgets_Template = [];
-          CacheService.allSamples = [];
-          CacheService.allWidgets_Sample = [];
-          navigate('RESTART APP');
-        },
-      },
-    ]
+      };
+      CacheService.allProjects = [];
+      CacheService.allWidgets_Project = [];
+      CacheService.allWidgets_Template = [];
+      CacheService.allSamples = [];
+      CacheService.allWidgets_Sample = [];
+      navigate('RESTART APP');
+    }
   );
 }