+++ docs/ReleaseNotes.md
@@ -1,3 +1,5 @@
 // Before any commit, use the command line: tsc --noEmit
 
-# 0.1.2.12 Language Tags refactor
+# 0.1.2.13 New Aptics Service
+
+- All expo aptics call replaced by the service call